Fix #41 : Remove unnecessary sort from map printing, closes #40
Map sort was there originally to ensure coherent print of MethodCall args.
It is unnecessary cause Groovy map literals create LinkedHashMap, which preserves insertion order.

ClassCastException was produced due to the sort operation which used a TreeMap, resulting in a
comparison of string w/ GString, which is problematic in Groovy.