Add Retriever for testing pipeline steps of a project itself.
The retrievers available up to now are designed for testing
a pipeline which makes use of pipeline steps. In the focus of
the test is a pipeline, not the pipeline steps provided by a
shared lib.

The retriever provided here helps with testing pipeline steps
provided by a shared lib project itself.