Update library loading so branch not required for localSource locations.
I was trying to use JenkinsPipelineUnit to test a library I'm writing.
I want to use this from within local copy (clone) of the repo using the
localSource retriever. This was problematic as the code required setting
defaultVersion (branch) which it then appended to the path to the
library root.

This PR allows the following usage:

```
        def library = library()
                .name('jenkins-library')
                .allowOverride(true)
                .implicit(false)
                .targetPath(dirPath)
                .retriever(localSource(dirPath))
                .build()
        helper.registerSharedLibrary(library)
```
where `dirPath` is the project root, without `@master` or other version
identifier in the filesystem path.