Make closure to registerAllowedMethod optional
Before, you had to specify the closure to registerAllowedMethod, even
if you wanted the default behaviour. For this reason, a big block of
registerAllowedMethod calls has a lot of noise, because each line ends
with `, null)`. This chang makes it possible to omit.

Technically, this change is not API-compatible, but I think the benefits
outweigh the small amount of corner-cases this might break...