Support not cloning method call arguments
v1.3 introduced a guard against call stack arguments mutating away from
the values they had at the time of the original method call. The guard
is to clone the argument instances at the time of the call.

In most use cases this ideal. But for arguments whose types are not
`Cloneable` the guard turns the arg values into `String` instances.
This is good fallback behavior but can break tests that do
fine-grained validations against nested structures, for example
`UnmodifiableMap` instances.

In those use cases the pre-v1.3 behavior of preserving the original
references may be desirable. The new
`PipelineTestHelper.cloneArgsOnMethodCallRegistration` property can
be set by the test to disable the cloning.