Allow for mocking sh step output
The sh step is a bit special, because it is the workhorse of many
pipeline scripts. In many cases, users need to be able to test how a
pipeline behaves if a script fails, or if it returns some output that
is required as the input for subsequent steps.

This commit adds a new method to PipelineTestHelper called addShMock,
which accepts either a combination of an exit code and return value,
or a closure. In both cases, when the mocked sh step runs, it looks to
see if any output has been registered for the given script. If it
finds a match, it returns either the exit code and/or return value
(depending on the arguments given to the sh step), or executes the
closure and returns that value.

Note that for this feature to work, the mocked script call must
EXACTLY match the call made to the sh step. In other words, globs or
fuzzy matching of the script call is not yet supported.

If the sh step does not find a match for the given script, then the
old behavior is retained and a hardcoded "bbb\nccc\n" value is
returned, or a fake SHA1 value if the script call happened to be "git
rev-parse HEAD".

At some point in the future, I would like to remove this behavior, but
I felt it important to add this feature without introducing a breaking
change in order to give users of this framework time to adapt their
test cases.