Allow for mocking readFile step output
This commit adds a new method to PipelineTestHelper called
addReadFileMock, which allows one to specify the contents which should
be returned by a call to readFile. This is useful for testing
pipelines which read files (perhaps generated by a build step, are
fetched from a remote server, etc), and the contents of these files
influence subsequent build logic.

If the readFile mock is called by a pipeline without the unit test
having called addReadFileMock, then an empty string is returned.