Fix IllegalArgumentException with Global Var taking a single argument
When calling a Gloval Var declared to accept one argument and the caller
passes null to the function it yields an IllegalArgumentException
instead of passing the null argument. This is unique to single argument
method calls and appears rooted in the way Groovy lang handles the
invocation when null is provided.

This fix was inspired from the code in the invoked Groovy methods.