Remove dummy values for unconfigured `sh()` and `bat()` mock calls.
Only return an empty string if stdout is requested. Under some conditions,
this used to return dummy values even if no output was requested, which is
clearly a bug.

Also, the backward compatibility code has been maintained there for long
enough for users to update their code if it depends on it.