Extend nested helper class.
Instead of just holding some data, encapsulate both matching a call and
executing a mock operation. Further, instead of a map, use a stack of
handlers.

This mostly reorders some code, but also prepares for patterned call
matching. As a first addition, it allows you to configure a fallback
handler (`pattern = null`) that matches everything. This is useful if
you want to detect unexpected calls.