Make bindings accessible in SharedLibrary classes.
For the library loading a Pogo is instantiated, which has a different implementation while getting an object.

This POGO does not read the bindings. In the PogoMetaClassGetPropertySite the implementation can be found. This throws a runtime exception, which suits well for the common purposes, but not when we try to emulate the Jenkins shared library behaviour.

So the workaround is to add the bindings as properties dynamically inside InterceptingGCL.