Don't use a single type in MockScriptHandler()
The previous implementation of the MockScriptHandler constructor was a
bit brittle and confusing. The first parameter was an Object which could
either be null (for no pattern matching), a String (for full pattern
matching), or a Pattern (for a regexp pattern match).

This change now provides three different constructors for the three
different use-cases detailed above, and also creates separate fields for
the various matches. This means we no longer need to rely on instanceof
to see what type of filter should be applied.