Remove TestNotSerrialzibleCPS and related files
This test never worked properly because it expected any Exception to be
thrown, and in fact, this test throws:

groovy.lang.MissingMethodException: No signature of method:
com.lesfurets.jenkins.TestNotSerrialzibleCPS.containsString() is
applicable for argument types: (java.lang.String) values: [Unable to
serialize locals]

This can be fixed by statically importing containsString, but when that
is done, the test passes because there is no Serializable exception
thrown by the test case.

Overall, I think that this test isn't contributing much to the framework
and it should instead be removed.