Allow declaration of kubernetes agent with no config
The kubernetes agent declaration supports using default/unspecified
pod (similar to "agent any") by passing true instead of specifying any
details. The pipeline syntax generator even provides this when you do
not set any options for it.

However, since AgentDeclaration's interface accepts an Object and
tries to cast it to a KubernetesAgentDeclaration, attempting to test a
pipeline using this syntax fails with a casting error since Boolean
cannot be cast to KubernetesAgentDeclaration. Therefore, add a
specialization of the method that accepts Boolean and creates an empty
KubernetesAgentDeclaration.